Crate rod

The Rust implementation of the GUN decentralized database syncronization protocol.

Rod is attempting to be compatible with the official JavaScript implementation of GUN while also supporting extra features such as a binary serialization

Modules

crdt
CRDTs ( Conflict-free Replicated Data Types ), including the implementation of GUN’s HAM algorithm

engine
Contains the main Rod struct, used to access the replicated database

executor
Functions abstracted over the current async executor such as spawn

graph
Data structures used to build the rod data graph

prelude
The Rod prelude

protocol
Wire protocol data types and representations

store
Backing data stores used for persistant data

tap
tap – Syntactical Plumb-Lines

Structs

Ulid
A Ulid is a unique 128-bit lexicographically sortable identifier