Trait rod::store::Store

pub trait Store {
    #[must_use]
    fn get_node<'life0, 'life1, 'async_trait>(
        &'life0 self, 
        id: &'life1 Ulid
    ) -> Pin<Box<dyn Future<Output = Result<Option<Node>, StoreError>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;

    #[must_use]
    fn put_node<'life0, 'async_trait>(
        &'life0 self, 
        node: Node
    ) -> Pin<Box<dyn Future<Output = Result<(), StoreError>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

    #[must_use]
    fn delete_node<'life0, 'life1, 'async_trait>(
        &'life0 self, 
        id: &'life1 Ulid
    ) -> Pin<Box<dyn Future<Output = Result<(), StoreError>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;

    #[must_use]
    fn set_id<'life0, 'life1, 'async_trait>(
        &'life0 self, 
        key: &'life1 str, 
        id: Option<Ulid>
    ) -> Pin<Box<dyn Future<Output = Result<(), StoreError>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;

    #[must_use]
    fn get_id<'life0, 'life1, 'async_trait>(
        &'life0 self, 
        key: &'life1 str
    ) -> Pin<Box<dyn Future<Output = Result<Option<Option<Ulid>>, StoreError>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;
}

A Node storage interface

Store is designed to be implemented over any persistant storage interface such as the filesystem, S3, browser LocalStorage or IndexedDB, etc. but is also responsible for any in-memory caching or buffering that may need to be done to optimize for performance.

Callers will not be expected to have to manually flush do disk or other permanant storage, but it will be the responsibility of the store.

Required methods

#[must_use]
fn get_node<'life0, 'life1, 'async_trait>(
&'life0 self,
id: &'life1 Ulid
) -> Pin<Box<dyn Future<Output = Result<Option<Node>, StoreError>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

Get a node from the store using it’s ULID

#[must_use]
fn put_node<'life0, 'async_trait>(
&'life0 self,
node: Node
) -> Pin<Box<dyn Future<Output = Result<(), StoreError>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Put a node into the store

The node can later be retrieved using it’s ULID

#[must_use]
fn delete_node<'life0, 'life1, 'async_trait>(
&'life0 self,
id: &'life1 Ulid
) -> Pin<Box<dyn Future<Output = Result<(), StoreError>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

Delete a node from the store using it’s ULID

#[must_use]
fn set_id<'life0, 'life1, 'async_trait>(
&'life0 self,
key: &'life1 str,
id: Option<Ulid>
) -> Pin<Box<dyn Future<Output = Result<(), StoreError>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

Point a string key in the database to a node’s ULID

#[must_use]
fn get_id<'life0, 'life1, 'async_trait>(
&'life0 self,
key: &'life1 str
) -> Pin<Box<dyn Future<Output = Result<Option<Option<Ulid>>, StoreError>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

Get the ULID pointed at by the string key in the database